+++ apps/web/src/views/Home/components/Banners/hooks/useMultipleBannerConfig.tsx
@@ -1,6 +1,5 @@
 import shuffle from 'lodash/shuffle'
 import { useMemo, type ReactElement } from 'react'
-import { BrasilMeetupBanner } from 'views/Home/components/Banners/BrasilMeetupBanner'
 import { AlloraBanner } from '../AlloraBanner'
 import { BirthdayBanner } from '../BirthdayBanner'
 import CompetitionBanner from '../CompetitionBanner'
@@ -50,10 +49,6 @@ export const useMultipleBannerConfig = () => {
       {
         shouldRender: true,
         banner: <WebNotificationBanner />,
-      },
-      {
-        shouldRender: true,
-        banner: <BrasilMeetupBanner />,
       },
       {
         shouldRender: true,