+++ packages/farms/constants/polygonZkEVMTestnet.ts
@@ -2,6 +2,7 @@ import { polygonZkEvmTestnetTokens } from '@pancakeswap/tokens'
 import { FeeAmount } from '@pancakeswap/v3-sdk'
 import { defineFarmV3Configs } from '../src/defineFarmV3Configs'
 
+/** @deprecated */
 export const farmsV3 = defineFarmV3Configs([
   {
     pid: 1,