+++ packages/farms/src/farms/base.ts
@@ -1,7 +1,8 @@
 import { ChainId } from '@pancakeswap/chains'
 import { baseTokens } from '@pancakeswap/tokens'
 import { FeeAmount, Pool } from '@pancakeswap/v3-sdk'
-import { Protocol, UniversalFarmConfig } from '../types'
+import { defineFarmV3ConfigsFromUniversalFarm } from '../defineFarmV3Configs'
+import { Protocol, SerializedFarmConfig, UniversalFarmConfig, UniversalFarmConfigV3 } from '../types'
 
 const pinnedFarmConfig: UniversalFarmConfig[] = [
   {
@@ -71,6 +72,15 @@ const pinnedFarmConfig: UniversalFarmConfig[] = [
 
 export const baseFarmConfig: UniversalFarmConfig[] = [
   ...pinnedFarmConfig,
+  {
+    pid: 39,
+    chainId: ChainId.BASE,
+    protocol: Protocol.V3,
+    lpAddress: Pool.getAddress(baseTokens.dola, baseTokens.usdc, FeeAmount.MEDIUM),
+    token0: baseTokens.dola,
+    token1: baseTokens.usdc,
+    feeAmount: FeeAmount.MEDIUM,
+  },
   {
     pid: 38,
     chainId: ChainId.BASE,
@@ -356,3 +366,11 @@ export const baseFarmConfig: UniversalFarmConfig[] = [
 ]
 
 export default baseFarmConfig
+
+/** @deprecated */
+export const legacyV3BaseFarmConfig = defineFarmV3ConfigsFromUniversalFarm(
+  baseFarmConfig.filter((farm): farm is UniversalFarmConfigV3 => farm.protocol === Protocol.V3),
+)
+
+/** @deprecated */
+export const legacyFarmConfig: SerializedFarmConfig[] = []