+++ packages/farms/src/farms/polygonZkEVM.ts
@@ -1,7 +1,8 @@
 import { ChainId } from '@pancakeswap/chains'
 import { polygonZkEvmTokens } from '@pancakeswap/tokens'
 import { FeeAmount, Pool } from '@pancakeswap/v3-sdk'
-import { Protocol, UniversalFarmConfig } from '../types'
+import { defineFarmV3ConfigsFromUniversalFarm } from '../defineFarmV3Configs'
+import { Protocol, SerializedFarmConfig, UniversalFarmConfig, UniversalFarmConfigV3 } from '../types'
 
 const pinnedFarmConfig: UniversalFarmConfig[] = [
   {
@@ -210,3 +211,11 @@ export const polygonZkEVMFarmConfig: UniversalFarmConfig[] = [
 ]
 
 export default polygonZkEVMFarmConfig
+
+/** @deprecated */
+export const legacyV3PolygonZkEVMFarmConfig = defineFarmV3ConfigsFromUniversalFarm(
+  polygonZkEVMFarmConfig.filter((farm): farm is UniversalFarmConfigV3 => farm.protocol === Protocol.V3),
+)
+
+/** @deprecated */
+export const legacyFarmConfig: SerializedFarmConfig[] = []