+++ packages/farms/src/farms/polygonZkEVMTestnet.ts
@@ -1,7 +1,8 @@
 import { ChainId } from '@pancakeswap/chains'
 import { polygonZkEvmTestnetTokens } from '@pancakeswap/tokens'
 import { FeeAmount, Pool } from '@pancakeswap/v3-sdk'
-import { Protocol, UniversalFarmConfig } from '../types'
+import { defineFarmV3ConfigsFromUniversalFarm } from '../defineFarmV3Configs'
+import { Protocol, SerializedFarmConfig, UniversalFarmConfig, UniversalFarmConfigV3 } from '../types'
 
 const pinnedFarmConfig: UniversalFarmConfig[] = []
 
@@ -37,3 +38,11 @@ export const polygonZkEVMTestnetFarmConfig: UniversalFarmConfig[] = [
 ]
 
 export default polygonZkEVMTestnetFarmConfig
+
+/** @deprecated */
+export const legacyV3PolygonZkEVMTestnetFarmConfig = defineFarmV3ConfigsFromUniversalFarm(
+  polygonZkEVMTestnetFarmConfig.filter((farm): farm is UniversalFarmConfigV3 => farm.protocol === Protocol.V3),
+)
+
+/** @deprecated */
+export const legacyFarmConfig: SerializedFarmConfig[] = []