+++ packages/gauges/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pancakeswap/gauges
 
+## 1.2.15
+
+### Patch Changes
+
+- 6d4ecf0: List farm, pool, gauges
+- Updated dependencies [6d4ecf0]
+  - @pancakeswap/tokens@0.6.25
+
 ## 1.2.14
 
 ### Patch Changes